Deduplicate server and client flight code
